feat-#16: move queries to application module
