+++ src/item/application/queries/handlers/get-item-by-id.handler.ts
@@ -1,10 +1,10 @@
 import { QueryHandler, IQueryHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { RpcExceptionService } from '../../../utils/exception-handling';
-import { ItemReadRepository } from '../../repositories';
+import { RpcExceptionService } from '../../../../utils/exception-handling';
+import { ItemReadRepository } from '../../../repositories';
 import { GetItemByIdQuery } from '../impl';
-import { ItemEntity } from '../../entities';
+import { ItemEntity } from '../../../entities';
 
 @QueryHandler(GetItemByIdQuery)
 export class GetItemByIdHandler implements IQueryHandler<GetItemByIdQuery> {
