+++ src/item/application/queries/handlers/get-items.handler.ts
@@ -1,9 +1,9 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { ItemEntity } from '../../entities';
-import { ItemReadRepository } from '../../repositories';
-import { RpcExceptionService } from '../../../utils';
+import { ItemEntity } from '../../../entities';
+import { ItemReadRepository } from '../../../repositories';
+import { RpcExceptionService } from '../../../../utils';
 import { GetItemsQuery } from '../impl';
 
 @QueryHandler(GetItemsQuery)
