+++ src/item/application/queries/impl/get-item-by-id.query.ts
