+++ src/item/application/queries/impl/get-items.query.ts
@@ -1,4 +1,4 @@
-import { GetItemsDto } from '../../dtos';
+import { GetItemsDto } from '../../../dtos';
 
 export class GetItemsQuery {
   constructor(public readonly getItemsDto: GetItemsDto) {}
